# backend/service-b/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Event logs and PDF reporting",
        "title": "Service B API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/health": {
            "get": {
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/logs": {
            "get": {
                "tags": [
                    "logs"
                ],
                "summary": "Query stored event logs",
                "parameters": [
                    {
                        "type": "string",
                        "default": "GET /search",
                        "description": "Filter by event type",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "2020-01-01T00:00:00.000Z",
                        "description": "ISO date from",
                        "name": "date_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "2030-12-31T23:59:59.999Z",
                        "description": "ISO date to",
                        "name": "date_to",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 200,
                        "minimum": 1,
                        "type": "integer",
                        "default": 50,
                        "description": "Page size",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_fs-assignment_service-b_internal_service.PaginatedLogs"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/report/pdf": {
            "get": {
                "tags": [
                    "report"
                ],
                "summary": "Generate a PDF report with charts via the Go gRPC service",
                "parameters": [
                    {
                        "type": "string",
                        "default": "2026-07-20",
                        "description": "Report day (UTC), defaults to today",
                        "name": "date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "all",
                        "description": "Filter by telemetry action label, or 'all'",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_fs-assignment_service-b_internal_models.LogEntry": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "eventType": {
                    "type": "string"
                },
                "payload": {
                    "type": "object",
                    "additionalProperties": true
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "github_com_fs-assignment_service-b_internal_service.PaginatedLogs": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_fs-assignment_service-b_internal_models.LogEntry"
                    }
                },
                "meta": {
                    "$ref": "#/definitions/github_com_fs-assignment_service-b_internal_service.PaginationMeta"
                }
            }
        },
        "github_com_fs-assignment_service-b_internal_service.PaginationMeta": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        }
    }
}